Move showLoading() to GameState
